+++ src/request-security.js
@@ -57,7 +57,7 @@ const CONTENT_SECURITY_POLICY=[
   "default-src 'self'",
   "script-src 'self'",
   "style-src 'self' 'unsafe-inline'",
-  "img-src 'self' data: blob:",
+  "img-src 'self' data: blob: https:",
   "font-src 'self'",
   "connect-src 'self'",
   "media-src 'self' blob:",